Convert most of SSO API surface to internal
