--- sso/src/main/java/dev/drewhamilton/skylight/sso/dagger/Sso.kt
@@ -1,6 +1,0 @@
-package dev.drewhamilton.skylight.sso.dagger
-
-import javax.inject.Qualifier
-
-@Qualifier
-internal annotation class Sso
