+++ sso/src/main/java/dev/drewhamilton/skylight/sso/network/SsoApi.kt
@@ -5,7 +5,7 @@ import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
 
-interface SsoApi {
+internal interface SsoApi {
 
     @GET("json")
     fun getInfo(
