+++ sso/src/main/java/dev/drewhamilton/skylight/sso/network/request/Params.kt
@@ -2,4 +2,4 @@ package dev.drewhamilton.skylight.sso.network.request
 
 import dev.drewhamilton.skylight.sso.network.SsoDate
 
-data class Params(val lat: Double, val lng: Double, @SsoDate val date: String)
+internal data class Params(val lat: Double, val lng: Double, @SsoDate val date: String)
