+++ sso/src/main/java/dev/drewhamilton/skylight/sso/network/response/SsoInfoResponse.kt
@@ -1,6 +1,6 @@
 package dev.drewhamilton.skylight.sso.network.response
 
-data class SsoInfoResponse(
+internal data class SsoInfoResponse(
     val results: SunriseSunsetInfo,
     val status: String
 )
