+++ sso/src/main/java/dev/drewhamilton/skylight/sso/network/response/SunriseSunsetInfo.kt
@@ -2,8 +2,7 @@ package dev.drewhamilton.skylight.sso.network.response
 
 import dev.drewhamilton.skylight.sso.network.SsoDateTime
 
-// TODO: Convert to DSL thing
-data class SunriseSunsetInfo(
+internal data class SunriseSunsetInfo(
     @SsoDateTime val sunrise: String,
     @SsoDateTime val sunset: String,
     @SsoDateTime val civil_twilight_begin: String,
